requirements.txt
aiofiles
aiosqlite
aniso8601
appdirs
async-exit-stack
async-generator
asyncpg
#backports.zoneinfo
black
cachetools
certifi
chardet
click
colorama
commonmark
dnspython
email-validator
fastapi
flake8
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
graphene
graphql-core
graphql-relay
h11
httplib2
httptools
idna
iso8601
itsdangerous
Jinja2
MarkupSafe
mccabe
mypy-extensions
oauthlib
orjson
packaging
pathspec
#pkg_resources
platformdirs
promise
protobuf
pyasn1
pyasn1-modules
pycodestyle
pydantic
pyflakes
Pygments
pyparsing
pypika-tortoise
python-dotenv
python-multipart
pytz
PyYAML
regex
requests
requests-oauthlib
rich
rsa
Rx
sentry-sdk
six
starlette
toml
tomli
tortoise-orm
typing-extensions
tzdata
ujson
uritemplate
urllib3
uvicorn
uvloop
watchgod
websockets